add reset state to middleweare
