+++ src/store/constants.js
@@ -12,3 +12,4 @@ export const UPDATE_QUOTES = 'UPDATE_QUOTES';
 export const UPDATE_STOCK_TICKER = 'UPDATE_STOCK_TICKER';
 export const UPDATE_CHART_DATA = 'UPDATE_CHART_DATA';
 export const UPDATE_SEARCH_QUERY = 'UPDATE_SEARCH_QUERY';
+export const RESET = 'RESET';
