+++ src/store/store.js
@@ -3,24 +3,9 @@ import { startupMiddleware, stockMiddleware } from './middleware'
 import {reducer} from './reducer'
 import { bootstrap } from './actions'
 
-const initialState = {
-    selectedStock: '',
-    selectedChartRange: '5y',
-    selectedKeyStats:'',
-    selectedLatestNews:null,
-    selectedCompanyOverview:'',
-    selectedTopPeers:null,
-    selectedSearch:'',
-    selectedCompanySymbols:[],
-    selectedQuotes:'',
-    selectedStockTicker:'',
-    selectedChartData:[],
-    enteredSearchQuery:'',
-}
-
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
-export const store = createStore(reducer, initialState,
+export const store = createStore(reducer, undefined,
     composeEnhancers(
         applyMiddleware(...[stockMiddleware, startupMiddleware])
     ))
